Fix chpldoc bug with single statement functions
This commit fixes a bug where the parser would drop a comment if the
preceding function was a single-statement body ending in an else-less
'if then'. For example:

```chpl
/* doc foo */
proc foo() do if false then ;

/* doc bar */
proc bar() { }
```

Here, the 'doc bar' docstring would be dropped entirely by the parser.

The reason for this was that when parsing the if-then, the parser would
scoop up the 'doc bar' comment while looking for the end of the if-then
statement. Then, the parser would proceed to clear its comments at
various points while building the function.

The solution is to instead invoke 'clearComments' with a set of comments
gathered from the last location represented by the function.

A test is added to lock in this behavior.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>