Adjust the Help documentation tests w.r.t. multilocale settings
The example codes involve printing out the arguments passed to main.  The first
argument passed to main is the program name, which is pretty simple in single
locale settings.  However, in multilocale settings this is the `_real` binary,
and as a result we get the full path to that binary (which will vary from system
to system).

While it is in general valuable to adjust tests so that they're robust in these
scenarios, I don't think it's appropriate to do so in this instance.  The
part that varies is not something done by the Help module itself, it's an
example for the user of what *they* could do.  With that in mind, I'm skipping
the tests in multilocale scenarios.  I've also added a note that the program
name will change to the test, so that readers will be alerted to this when
reading the example code.

Validated that the tests are skipped in multilocale settings now

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>