Update the compiler to use `preview` instead of `pre-edition` and warn
We decided to rename the edition where we accumulate changes from `pre-edition`
to `preview`.  Update the name stored in the editions array, but continue to
accept `pre-edition` for a while longer, generating a warning when the old name
is used.

Unfortunately, this warning will fire three times instead of just once.  I
believe this is because we do several subprocesses for the compiler and flags
need to be checked at all of them.  I tried several strategies to make the
message get printed just once but I think the only way to truly manage it is to
make it an error instead of a warning and I don't think that's appropriate

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>