Add skipif for settings where this test doesn't behave the same
The vectorization output doesn't always get generated, so we don't want to run
the test when it would generate different output.  Since it gets output in the
default settings, I think it's okay to just skipif this test instead of
splitting the LLVM portion into its own test (but that's probably the more
principled solution).

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>