Adjust the assertVectorized specific test to not have an empty body
There was a question about whether it should warn about not vectorizing in this
case, and the simplest way to resolve that question is to give it an example
body since no user code is going to make the loop body empty.  In this case, I
gave it an example that should always vectorize (and made the docs explicit
about that expectation).

Removed the warning message from the expected output with that in mind

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>