Improve default argument generation
This commit updates ``convertActual`` to always generate a symbol for
the actual, which may require a temporary. This allows us to simplify
the implementation of ``interceptFormalUse``, and correctly generate
default argument values that reference other formals.

Also updates ``interceptFormalUse`` to use the correct actualState_
index.

A test is added to ensure this continues to work.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>