Refactor actual conversion
This commit refactors conversion of actuals to avoid relying on the
actual index or formal index to refer to certain stored data. Before, we
were doing this with 'actualState_', but the actualIdx and formalIdx can
sometimes conflict with each other.

Instead, we get rid of 'actualState_' altogether. We now append
converted actuals to the desired call immediately after conversion.
Also, instead of a map from ID to FormalActual, we just store a map from
ID to the actual's ``Symbol*``.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>