Make progress towards supporting procedure pointers with C backend
Remove some old code doing GC of 'FunctionType'. It's not doing what
I thought it did and could cause 'null' holes in the 'gTypeSymbols'
global vector. I'll worry about GC with '--verify' at a later point
in this effort.

There is context-sensitive code in 'SymExpr::codegen()' which checks
to see how a 'SymExpr' over a 'FnSymbol' is used. If it looks like
it's being used as value, we generate it differently by adding it to
the procedure pointer cache and producing a wide index.

Previously this code was only firing for LLVM. I moved it out into a
more "backend agnostic" block of code.

I added code to generate a C function type from a 'FunctionType', but
currently that code is commented out. I am opting to generate 'wide
proc' as 'int64_t' and 'local proc' as 'void*' in order to work around
circular dependencies in C (e.g., a record contains a procedure
pointer and the procedure pointer accepts that record as a formal).

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>