Add support for using globals declared in internal/standard modules
In the typed converter we will increasingly encounter situations where
we need to resolve some global variable in an internal or standard
module, but are unable to resolve the entirety of that module. This
commit adds support for resolving just the variables that are used, by
making a copy of that variable for the typed converter's usage. The
original production AST is still generated.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>