Remove package support for Debian 11
I've been working on dropping support for Python 3.9 so we can update some of
our Python dependencies.  Debian 11 relies on Python 3.9, so that means we need
to stop supporting a package for it (though people can still build on that
machine by hand, and will even still be able to build chpldoc if they've
manually updated the version of Python they have present).

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>