Remove Red Hat Enterprise Linux 9 support
Similar to the previous commit - RHEL 9 relies on Python 3.9, but we are
dropping support for Python 3.9 with chpldoc, so can no longer provide a package
for that platform.  Chapel itself should still build fine there, and chpldoc
itself should still build fine with an updated version of Python on that
platform.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>