Add working test of loading Chpl lib dynamically; plus future
This adds a working test of creating and loading a simple
Chapel library (which is to say, one that doesn't rely on
the runtime) and making a call to it.  It also adds a future
for a case that seemed incorrect in which loading a non-existent
routine seemed to generate an assertion error rather than a
thrown error, as documented (at least on Mac).

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>