Add start of multilocale support; fix skipifs for Mac
This changes callAddEm into a multilocale test that works if compiling
the dynamic library as a single-locale library; but that's not something
we can reliably do within the test system (the runtime is unlikely to
have been built for single-locale compiles), so this wont actually
be tested currently.  It does work in a config that has both configs
built, though, by adding --comm=none --launcher=none to the PRECOMP
file's compiler options.

callAddEm-normalexit works on Mac, so adjust the skipifs so that:

* directory-wide SKIPIF skips for CHPL_COMM!=none
* per-test .skipifs pay attention to Linux or not and PIC, where
  - "normalexit" version skips if not on Linux or PIC is not set
  - others skip if on linux and PIC is not set

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>