helpConvertFields should preserve type/param fields
We'll need to keep the type/param fields around temporarily to re-use
some existing code related to instantiation. Also, make
'helpConvertFields' sensitive to whether the type is an instantiation or
not. If it isn't an instantiation, we should preserve the initialization
and type expressions of a field in case the production resolver needs to
use the type.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>