Add the edition attribute to the attributes technote
The mention is brief because the attribute is documented in more detail in the
already existing developer documentation on it, so it links to the relevant
documents.

While here, I adjusted the section link and references to it to use the current
section name.  I'd be open to splitting the chpldoc attribute into its own
Tooling Attributes section instead.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>