Make 'out' handling more robust
Instead of (somewhat hackily) making 'out' formals
count as "concrete" for the purposes of function
instantiation, consider them generic, and distinguish
between two types of genericity for function signatures.
This is important because a signature can be generic
(due to an 'out' formal that's generic) while still
being safe to resolve. Indeed, the only way to instantiate
a generic 'out' formal is from resolving the function's body.

Thus, with this commit, we track the two orthogonal
"genericities" (out formal genericity, and other genericity).

This PR also makes an adjustment in which an 'out' array type
expressions are counted as concrete, instead of generic
as they are counted elsewhere in function signatures.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>