Dyno: avoid asserting cases in call-init-deinit
The call-init-deinit code assumes that move initialization
is called with known LHS and RHS. In some error-producing cases,
an out formal's type can be generic or unknown, which results
in violating this assumption.

I adjusted the surrounding logic which calls resolveMoveInit,
instead of weakining the assumption, because it seems attractive
to let that assertion catch other cases of unknown LHS/RHS
for us to inspect and decide what to do with.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>