Adjust tuple type indexing to support types stored in fields
Previously, the logic that rewrote tuple access into .this
was only performed on variables. This included type variables,
but type fields are TypeSymbols (still with the TYPE_VARIABLE
flag). This PR adjusted the logic to also convert TYPE_VARIABLE
field calls into 'this' calls.

Note that there's an existing exception that doesn't convert
type calls into 'this' for all types except tuples. This is
presumably to enable invoking type constructors (as opposed
to the `this` type-level method). I kept this exception,
so you still can't call 'this' on non-tuple type fields.
This is also the case for non-tuple type variables, though,
so I'm not upset by it.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>