Fix issues with 'in'-intents when copying changes type
Task functions' 'in' intents are resolved differently from
regular 'in' intents. The compiler computes
'copyTypeDuringResolution' for normal generic 'in' formals
(thereby adjusting the functions to accept default arrays
even when the argument is an array view, because copying
an array view creates a new array). However, this doesn't
happen in task functions. As a result, we end up in a world
in which the generated function expects an array slice,
but gets an array. This error is uncaught at resolution
(presumably due to the special process of resolving task
function calls) and leads instead to segmentation faults.

This PR fixes the issue by invoking copyTypeDuringResolution
while configuring task function formals, mirroring the
regular 'in' formal case.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>