Inline the non-param nudge upper/lower bound routines
While looking at some generated code that used lo..<hi ranges today, I
was surprised to find a function call to compute the "hi-1" implied by
"hi", which seems like overkill in the generated C code.  This changes
the non-param versions of the routine to be inlined which seems like
it can only help, and likely not hurt, the compiler back-end.

---
Signed-off-by: Brad Chamberlain <bradcray@users.noreply.github.com>