Create patch to apply in dyno testing to ignore known errors
We'd like to silence a category of testing failures under '--dyno' when
we know that the test is not intended to compile. This helps to focus
our development cases on tests that are expected to compile
successfully.

The .patch file will be applied prior to our internal testing with the
'--dyno' flag enabled.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>