Turn candidate copy elisions into mentions if elision is not allowed
Prior to this change, if an AST node didn't allow copy elision, we simply
ignored all references variables in its body. This meant that
if a variable was mentioned in non-copy-elision-eligible block,
this mention would be ignored, and the variable would be considered
dead. This change fixes the issue by turning elisions into mentions
when an AST node doesn't support elision. Thus, ineligile AST is
still considered a mention, as opposed to being completely ignored.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>