Avoid inserting 'null' into array substitutions
We can construct an uninstanced array type that depends on
skipped expressions, in which case their type is 'QualifiedType()',
which has no pointer. However, `isInstantiation' logic for composite
types (which arrays are a special case of) eventually uses 'canPass'
on each substitution, and 'canPass' expects a non-null 'Type*'.
Thus, when we are feeding something directly into substitutions
by creating an uninstanced array type, ensure it's not null.

Signed-off-by: Danila Fedorin <daniel.fedorin@hpe.com>