Remove 'defer' blocks from lowered manage statements, use 'try!'
Previously I tried to lower 'manage' with 'defer' in order to make
sure that errors cleaned up on all paths. However this caused huge
problems when trying to wrap the 'manage' in a try/catch or interact
with throwing procedures in general. The reason I did not run into
this effort earlier in my own code is that I generally steer clear
of throwing context managers.

This is the first step in the effort to rectify that issue and make
'manage' behave properly in throwing contexts.

Instead of using 'defer', use a temp flag and call 'exitContext()' in
all paths of the lowered try/catch representing the manage statement.

In order to keep the behavior the same in a non-throwing context,
wrap the entire statement in a 'try!' block in order to make it so
that users do not have to wrap every 'manage' in a try/catch.

The next step is to adjust 'lowerErrorHandling()' in order to unwrap
the 'try!' so that things behave like normal in a throwing context.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>