Store arkouda HTML output alongside .dat files
This commit uses ``$CHPL_TEST_PERF_DESCRIPTION`` to unique-ify the
output HTML directories instead of using the same common directory
that happens to serve as a parent of the already unique .dat file
directories.

This manually mimics some behavior in arkouda's run_benchmarks script,
which already appends the description to the .dat file directory path.

The arkouda sub_test is updated with the new path for ``--graph-dir``,
and the common arkouda cron script is updated to rsync from those
directories.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>