Revert to old way of lowering manager, but introduce 'throwing defer'
This is a special variation of 'defer' in which the body can throw.
Normally, defers cannot throw as the contents must be atomic. But I
need a way to make sure the call to 'exitContext' happens on all
paths out of the 'manage' body (e.g., one path which I could not
handle with the previous lowered AST was a 'return' statement).

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>