Split 'resolveFieldDecl` to store results separately
This commit splits out some of the logic in ``resolveFieldDecl`` so that
we can re-query and fetch the RRBPID later. This new query is called
``resolveFieldResults``, and returns a type introduced by this commit
called ``ResolvedFieldResults``. This new type primarily stores the
RRBPID computed by a Resolver, but also stores some useful information
alongside the resolution results.

A convenience query, called ``resolvedFieldsFromResults``, that can be
combined with ``resolveFieldResults`` to achieve the same output as
``resolveFieldDecl``.

This commit also fixes a minor bug when initializing the ``ResolvedFields``
structure. Previously the given RRBPID was not const, so invoking
'byAst' would create an empty entry. Instead, we check if we're in
``syntaxOnly`` mode, in which case there will not be results and we
should instead create an unknown QualifiedType.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>