Store implicit initialization information in ResolvedFunction
This commit adds key data about implicitly initialized fields within an
initializer, to be used later during typed conversion. Initially this
information is a basic map of IDs before which the fields should be
initialized, but will likely need to evolve to handle more complex cases
related to e.g., branching. Currently, an empty ID indicates that the
fields need to be initialized at the end of the initializer.

InitResolver is udpated to collect this information, which is then
requested upon creation of the ResolvedFunction.

Signed-off-by: Ben Harshbarger <ben.harshb@gmail.com>