Clean up how 'FunctionType' are code generated, stop discarding them
I don't know if this code got mashed up due to excessive rebasing or
if I actually wrote it that way and was happy with it, but it's
nonsense that is causing problems. Anyhow, to fix a bug with nightly
testing we shouldn't be pruning function types at all, so stop doing
that. While here, rewrite the logic to collect function types to stop
being so nonsensical.

Signed-off-by: David Longnecker <dlongnecke-cray@users.noreply.github.com>